Surface sitemapper 'lastmod' flag
The library 'sitemapper' is used to parse sitemaps when configured to be included
for a crawl.  This library, since v3.2.5, supports a 'lastmod' option that will
filter the returned URLs to only those with a modified date less than or equal
to the timestamp passed.

This commit surfaces that library flag to the crawl configurations, allowing
the inclusion of URLs scoped from sitemaps to be limited by their last
modified date.

https://github.com/webrecorder/browsertrix-crawler/issues/357